Query view in MySQL results in collation mismatch error
When quering a view and applying a where clause eg.
`WHERE type = 'member'` MySQL returns the following error:

SQLSTATE[HY000]: General error: 1267 Illegal mix of collations
(utf8mb4_unicode_ci, COERCIBLE) and (utf8mb4_general_ci, COERCIBLE) for operation '='

It is suspected that because the entity is not a proper table it
could not correctly detect the appropiate collation to use.

Setting it explicitly resolves the issue.