fix: sync unable to detect login
Sync job was unable to login successfully because during login it was encountering a timeout.

After investigation it was found that the cookie (PHPSESSID) was changing half way through the process.
The current http client is stateless and has no concept of a cookie jar or share cookie session.
At least during the login process we check for a change to the cookie value and update it before continuing with the next request.